@@ -1,45 +1,49 @@
 {
-  "name": "action-gh-release",
-  "version": "0.1.14",
-  "private": true,
+  "name": "github-release-action",
   "description": "GitHub Action for creating GitHub Releases",
   "main": "lib/main.js",
   "scripts": {
-    "build": "ncc build src/main.ts --minify",
-    "test": "jest",
-    "fmt": "prettier --write \"src/**/*.ts\" \"__tests__/**/*.ts\"",
-    "fmtcheck": "prettier --check \"src/**/*.ts\" \"__tests__/**/*.ts\""
+    "build": "ncc build src/main.ts --source-map --minify --license licenses.txt",
+    "lint": "eslint src/**/*.ts __tests__/**/*.ts",
+    "format": "eslint --fix src/**/*.ts __tests__/**/*.ts",
+    "test": "jest --coverage",
+    "all": "yarn run build && yarn run format && yarn test"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/softprops/action-gh-release.git"
+    "url": "git+https://github.com/crazy-max/ghaction-github-release.git"
   },
   "keywords": [
-    "actions"
+    "actions",
+    "github",
+    "release"
   ],
-  "author": "softprops",
+  "author": "CrazyMax",
   "license": "MIT",
   "dependencies": {
     "@actions/core": "^1.4.0",
     "@actions/github": "^5.0.0",
-    "@octokit/plugin-retry": "^3.0.9",
-    "@octokit/plugin-throttling": "^3.5.1",
-    "glob": "^7.1.6",
-    "mime": "^2.4.4",
-    "node-fetch": "^2.6.1"
+    "glob": "^8.0.3",
+    "mime": "^3.0.0",
+    "node-fetch": "^3.3.0"
   },
   "devDependencies": {
-    "@types/glob": "^7.1.1",
-    "@types/jest": "^24.0.25",
-    "@types/mime": "^2.0.1",
-    "@types/node": "^12.12.24",
+    "@types/glob": "^8.0.0",
+    "@types/mime": "^3.0.1",
+    "@types/node": "^18.11.3",
     "@types/node-fetch": "^2.5.12",
-    "@zeit/ncc": "^0.21.0",
-    "jest": "^24.9.0",
-    "jest-circus": "^24.9.0",
-    "prettier": "1.19.1",
-    "ts-jest": "^24.2.0",
-    "typescript": "^3.7.4",
-    "typescript-formatter": "^7.2.2"
+    "@typescript-eslint/eslint-plugin": "^5.40.1",
+    "@typescript-eslint/parser": "^5.40.1",
+    "@vercel/ncc": "^0.34.0",
+    "dotenv": "^16.0.3",
+    "eslint": "^8.26.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-jest": "^27.1.3",
+    "eslint-plugin-prettier": "^4.2.1",
+    "jest": "^29.2.1",
+    "prettier": "^2.7.1",
+    "ts-jest": "^29.0.3",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.4"
   }
 }
